Add cloudinary delete feature

+++ controllers/posts.js
@@ -47,30 +47,53 @@ module.exports = {
 	async updatePost(req,res,next){
 		
 		let foundPost = await Post.findById(req.params.id);
-		let images = []
-		for(let x of foundPost.image){
-			images.push(x.public_id)
+		
+		if(req.body.pictures && req.body.pictures.length){
+			let images = []
+			for(let image of req.body.pictures){
+				images.push(image)
+				for(let img of foundPost.image){
+					if(img.public_id === image){
+						let index = foundPost.image.indexOf(img);
+						foundPost.image.splice(index, 1)
+					}				
+				}
+			}
+			await cloudinary.api.delete_resources(images);
+			
 		}
 		
-		await cloudinary.api.delete_resources(images);
-		req.body.image = []
-		for(let file of req.files){
-			let result = await cloudinary.uploader.upload(file.path)
-			req.body.image.push({
-				url: result.secure_url,
-				public_id: result.public_id
-			}) 
+		if(req.files){
+			for(let file of req.files){
+				let result = await cloudinary.uploader.upload(file.path)
+				foundPost.image.push({
+					url: result.secure_url,
+					public_id: result.public_id
+				}) 
+			}
+			
 		}
-		let updatePost = await Post.findByIdAndUpdate(req.params.id, req.body, {new: true})
-		res.redirect(`/posts/${updatePost.id}`)
 		
+		let {title, description, price, location } = req.body
+		foundPost.title = title
+		foundPost.description = description
+		foundPost.price = price
+		foundPost.location = location
+		let updatedPost = await foundPost.save({new:true})
+			
+		res.redirect(`/posts/${updatedPost.id}`)
 		// 	what it use to be	
 		// let update = await Post.findByIdAndUpdate(req.params.id, req.body, {new: true})
 		// res.redirect(`/posts/${update.id}`);
 	},
 	
 	async postDelete(req,res,next){
-		await Post.findByIdAndDelete(req.params.id);
+		let post = await Post.findById(req.params.id)
+		for(let image of post.image){
+			await cloudinary.uploader.destroy(image.public_id)
+		}
+		await post.remove();
+		// await Post.findByIdAndDelete(req.params.id);
 		res.redirect('/posts');
 	}
 }
